Django/cypos/requirements.txt
django-bootstrap3==7.0.0